refactor lifetime deal Hero component into separate HeroContent and PricingCard components, add appCount prop to agency partner HeroSection, fetch app count data in agency partner page, and add top margin to BlogGrid container || 1
